requirements.txt
astroid==2.2.5
entrypoints==0.3
flake8==3.7.7
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
pep8==1.7.1
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
python-decouple==3.1
python-http-client==3.1.0
sendgrid==6.0.5
six==1.12.0
typed-ast==1.3.5
wrapt==1.11.1
